refactor: 구현한 컴포넌트 디렉토리 변경
