+++ src/components/InputBox.js
@@ -2,15 +2,12 @@ import React from 'react';
 import styled from 'styled-components';
 
 const InputBoxStyle = styled.div`
-  background-color: #ecebf1;
-  height: 45px;
-  width: 100%;
-  text-align: center;
-  outline: 2px solid transparent;
-  outline-offset: 2px;
-  border-color: #9ca3af;
-  border: none;
+  display: flex;
+  align-items: center;
+  margin-top: 0.375rem;
+  color: #d3d3d3;
   border-radius: 0.25rem;
+  background-color: #ecebf1;
 `;
 
 export const InputBox = ({ children }) => {
